# Allow models to include methods for MiqExpression sql evaluation