This is currently a first pass at making it so models can define methods for SQL fragments for specific expression types.  Specifically, this allows the "INCLUDE ANY" to work in SQL for following expression:

```
Virtual Machine : IP Addresses INCLUDES ANY ['10.XXX.X']
```

Also of note:  The ruby implementation of `INCLUDES ANY` doesn't seem work as expected, so not sure if this is a bug with `MiqExpression` or I have implemented this expression in SQL incorrectly.


Benchmarks
----------
So there is some additional slowness issues that I noticed on master that don't seem to be present on the `gaprindashvili` branch, but they seem to be mitigated by this branch.  That said, seems like they are `LEFT JOIN` based issues, and could be an issue when requesting more that 20 Vms per-page, so it might be worth looking into at a later date.

For now, as show in the benchmarks, the times are WAY faster (and this isn't including another 200k ms request that also was slow on master, but now isn't an issue), so I think this fix is fine on it's own.

**BEFORE**

```
/vm_infra/report_data
|     ms | queries | query (ms) |   rows |
|   ---: |    ---: |       ---: |   ---: |
| 235682 |   22810 |    27690.0 | 278708 |
| 247473 |   22810 |    32461.1 | 278708 |
| 252947 |   22810 |    41896.3 | 278708 |

:rows_by_class:
  User: 2
  MiqGroup: 2
  Tenant: 2
  MiqUserRole: 2
  Entitlement: 1
  ManageIQ::Providers::InfraManager::Vm: 241115
  Network: 37584
:total_queries: 22810
:total_rows: 278708
```


**AFTER**

```
/vm_infra/report_data
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
|  857 |       9 |      417.5 |  209 |
|  525 |       9 |      328.1 |  209 |
|  510 |       9 |      316.6 |  209 |
|  515 |       9 |      308.7 |  209 |
|  460 |       9 |      289.8 |  209 |

:rows_by_class:
  User: 2
  MiqGroup: 2
  Tenant: 2
  MiqUserRole: 2
  Entitlement: 1
  ManageIQ::Providers::InfraManager::Vm: 200
:total_queries: 9
:total_rows: 209
```


Links
-----

* https://bugzilla.redhat.com/show_bug.cgi?id=1588082


Steps for Testing/QA
--------------------

1. Go to `Compute` > `Infrastructure` > `VirtualMachines` 
  
2. Select "All VMs" from the tree on the left
  
3. Click the down arrow to the right of the search box
  
4. Create the following query (replace "XXX.XXX.XXX" with something that makes sense for your environment)
  
  ```
  Virtual Machine : IP Addresses INCLUDES ALL ['XXX.XXX.XXX']
  ```

On master, this should be quite slow on a environment with a reasonable number of VMs (10k+), but very quick with this patch.