@jrafanie 2018-06-12
@NickLaMuro Looks good to me.  Can you fix any relevant cop complaints and add the BZ to the commit messages? Finally, do you have numbers for number of vms, hardwares, networks, etc. and time before/after this change?