@jrafanie 2018-06-12
> Doesn't work with loaded networks though... might not want to merge
this.

@NickLaMuro do you have an understanding of when networks is loaded (and how frequently) and can't use this optimization?  Is it an actual problem?