@NickLaMuro 2018-06-12
> > Doesn't work with loaded networks though... might not want to merge
this.

> @NickLaMuro do you have an understanding of when networks is loaded (and how frequently) and can't use this optimization? Is it an actual problem?

@jrafanie sorry, that commit message wasn't updated from when I only had the `else` portion of the current code change to `@ipaddresses ||=` in (see the diff snippet from just above [Keenan's comment here](https://github.com/ManageIQ/manageiq/pull/17562#discussion_r194423163)).  This is now not a relevant comment, so I will rebase it out.