@NickLaMuro 2018-06-12
> Can you fix any relevant cop complaints...

@jrafanie For this cop:

> **app/models/vm_or_template.rb**
> 
>   ❗️ - [Line 1645](https://github.com/NickLaMuro/manageiq/blob/f2785d0e9ba4c22ad8cee32e2ecfd7f3c05648cd/app/models/vm_or_template.rb#L1645), Col 24 - Layout/MultilineMethodCallIndentation - Align .or with .matches on line 1644.

I am tempted leave it as is.  Part of what I was trying to do was make the `arel` feel more like a SQL statement.  So how I wrote it out in the comments:

```SQL
1 = (SELECT 1
     FROM hardwares
     JOIN networks ON networks.hardware_id = hardwares.id
     WHERE hardwares.vm_or_template_id = vms.id
       AND (networks.ipaddress LIKE "%IPADDRESS%"
            OR networks.ipv6address LIKE "%IPADDRESS%")
     LIMIT 1
    )
```

So with the `networks` part of the `WHERE`, I tried to match it as best I could.  It got a bit gross keeping it as chained arel statement, so I broke it up a bit to make it a little easier and less "`rubocop` offending" (Murphy is easily offended).  The `.or` still doesn't fit the grade, but I think aligning it with `.matches` is harder to digest in this particular case.

> **lib/miq_expression.rb**
> 
>   ❗️ - [Line 1356](https://github.com/NickLaMuro/manageiq/blob/f2785d0e9ba4c22ad8cee32e2ecfd7f3c05648cd/lib/miq_expression.rb#L1356), Col 13 - Layout/ExtraSpacing - Unnecessary spacing detected.
>   ❗️ - [Line 1356](https://github.com/NickLaMuro/manageiq/blob/f2785d0e9ba4c22ad8cee32e2ecfd7f3c05648cd/lib/miq_expression.rb#L1356), Col 15 - Layout/SpaceAroundOperators - Operator = should be surrounded by a single space.

This is just one issue, duplicated to two cops, and while I honestly think it looks better with the extra spacing, it isn't worth arguing over, so I will make the change with my addition of the BZ link in the commit message and other rebasing tasks... you slavedriver...