@NickLaMuro 2018-06-12
> Finally, do you have numbers for number of vms, hardwares, networks, etc. and time before/after this change?

Yeah, I will add some benchmarking numbers after I do the rebase.