@NickLaMuro 2018-06-12
@jrafanie benchmarks added to the PR description.