@simaishi 2018-06-14
Gaprindashvili backport details:

```
$ git log -1
commit 8db5dcede0158678e71f80489e95e3d46860aa66
Author: Joe Rafaniello <jrafanie@users.noreply.github.com>
Date:   Tue Jun 12 16:39:47 2018 -0400

    Merge pull request #17562 from NickLaMuro/miq_expression_extra_includes_exp_sql_support
    
    Allow models to include methods for MiqExpression sql evaluation
    (cherry picked from commit 52176a7565126734cd960b9d3ed349e683ddb263)
    
    Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1591422
```